backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
google-generativeai==0.4.0
python-multipart==0.0.9
python-dotenv==1.0.1
pillow==12.1.1
sqlalchemy>=2.0.0
PyJWT>=2.8.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
sentence-transformers>=2.5.1
faiss-cpu>=1.8.0
numpy<2.0.0